wip add tests
